# Conflicts:
#	.github/workflows/ci.yml
#	symphonia-bundle-flac/Cargo.toml
#	symphonia-bundle-flac/src/demuxer.rs
#	symphonia-bundle-mp3/Cargo.toml
#	symphonia-bundle-mp3/src/demuxer.rs
#	symphonia-check/Cargo.toml
#	symphonia-codec-aac/Cargo.toml
#	symphonia-codec-aac/src/aac/mod.rs
#	symphonia-codec-aac/src/adts.rs
#	symphonia-codec-adpcm/Cargo.toml
#	symphonia-codec-alac/Cargo.toml
#	symphonia-codec-opus/Cargo.toml
#	symphonia-codec-pcm/Cargo.toml
#	symphonia-codec-vorbis/Cargo.toml
#	symphonia-codec-vorbis/src/codebook.rs
#	symphonia-codec-wavpack/Cargo.toml
#	symphonia-common/LICENSE
#	symphonia-core/Cargo.toml
#	symphonia-core/src/audio.rs
#	symphonia-core/src/codecs.rs
#	symphonia-core/src/formats.rs
#	symphonia-core/src/probe.rs
#	symphonia-format-caf/Cargo.toml
#	symphonia-format-isomp4/Cargo.toml
#	symphonia-format-isomp4/src/atoms/esds.rs
#	symphonia-format-isomp4/src/atoms/flac.rs
#	symphonia-format-isomp4/src/atoms/ilst.rs
#	symphonia-format-isomp4/src/atoms/mod.rs
#	symphonia-format-isomp4/src/atoms/opus.rs
#	symphonia-format-isomp4/src/atoms/stsd.rs
#	symphonia-format-isomp4/src/demuxer.rs
#	symphonia-format-mkv/Cargo.toml
#	symphonia-format-mkv/src/demuxer.rs
#	symphonia-format-mkv/src/ebml.rs
#	symphonia-format-ogg/Cargo.toml
#	symphonia-format-ogg/src/logical.rs
#	symphonia-format-riff/Cargo.toml
#	symphonia-format-riff/src/aiff/chunks.rs
#	symphonia-format-wav/Cargo.toml
#	symphonia-metadata/Cargo.toml
#	symphonia-play/Cargo.toml
#	symphonia-utils-xiph/Cargo.toml
#	symphonia/Cargo.toml
#	symphonia/src/lib.rs

+++ symphonia-bundle-mp3/src/header.rs
@@ -5,7 +5,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-use symphonia_core::errors::{decode_error, unsupported_error, Result};
+use symphonia_core::errors::{Result, decode_error, unsupported_error};
 use symphonia_core::io::ReadBytes;
 
 use crate::common::*;
@@ -14,7 +14,7 @@ use crate::common::*;
 pub const MPEG_HEADER_LEN: usize = 4;
 
 /// The maximum length in bytes of a MPEG audio frame including the header.
-pub const MAX_MPEG_FRAME_SIZE: u64 = 2881;
+pub const MAX_MPEG_FRAME_SIZE: usize = 2881;
 
 /// Bit-rate lookup table for MPEG version 1 layer 1.
 const BIT_RATES_MPEG1_L1: [u32; 15] = [
